Corrige une erreur lors de la mise en bêta avec un bot parmi les auteurs
