+++ zds/tutorialv2/tests/tests_views/tests_content.py
@@ -13,7 +13,6 @@ from django.contrib.auth.models import Group
 from django.http import HttpResponseNotAllowed
 from django.urls import reverse
 from django.test import TestCase
-from django.utils.translation import gettext_lazy as _
 
 from zds.forum.tests.factories import ForumFactory, ForumCategoryFactory
 from zds.forum.models import Topic, Post, TopicRead
@@ -23,7 +22,6 @@ from zds.gallery.models import Image
 from zds.member.tests.factories import ProfileFactory, StaffProfileFactory, UserFactory
 from zds.mp.models import PrivateTopic, PrivatePost
 from zds.notification.models import (
-    TopicAnswerSubscription,
     ContentReactionAnswerSubscription,
     NewPublicationSubscription,
     Notification,
@@ -49,7 +47,6 @@ from zds.tutorialv2.publication_utils import (
     ZMarkdownEpubPublicator,
 )
 from zds.tutorialv2.tests import TutorialTestMixin, override_for_contents
-from zds.utils.models import Tag
 from zds.tutorialv2.models.help_requests import HelpWriting
 from zds.utils.tests.factories import SubCategoryFactory, LicenceFactory
 from zds import json_handler
@@ -629,260 +626,6 @@ class ContentTests(TutorialTestMixin, TestCase):
         self.assertEqual(result.status_code, 200)
         beta_topic = Topic.objects.get(pk=beta_topic.pk)
         self.assertTrue(beta_topic.is_locked)
-
-    @patch("zds.tutorialv2.signals.beta_management")
-    def test_beta_workflow(self, beta_management):
-        """Test beta workflow (access and update)"""
-
-        # login with guest and test the non-access
-        self.client.force_login(self.user_guest)
-
-        result = self.client.get(reverse("content:view", args=[self.tuto.pk, self.tuto.slug]), follow=False)
-        self.assertEqual(result.status_code, 403)  # (get 403 since he is not part of the authors)
-
-        # login with author
-        self.client.force_login(self.user_author)
-        sometag = Tag(title="randomizeit")
-        sometag.save()
-        self.tuto.tags.add(sometag)
-        # create second author and add to tuto
-        second_author = ProfileFactory().user
-        self.tuto.authors.add(second_author)
-        self.tuto.save()
-
-        # activ beta:
-        tuto = PublishableContent.objects.get(pk=self.tuto.pk)
-        current_sha_beta = tuto.sha_draft
-        result = self.client.post(
-            reverse("content:set-beta", kwargs={"pk": tuto.pk, "slug": tuto.slug}),
-            {"version": current_sha_beta},
-            follow=False,
-        )
-        tuto = PublishableContent.objects.get(pk=self.tuto.pk)
-        self.assertEqual(result.status_code, 302)
-
-        # check if there is a pm corresponding to the tutorial in beta
-        self.assertEqual(Topic.objects.filter(forum=self.beta_forum).count(), 1)
-        self.assertTrue(PublishableContent.objects.get(pk=self.tuto.pk).beta_topic is not None)
-        self.assertEqual(PrivateTopic.objects.filter(author=self.user_author).count(), 1)
-        # check if there is a new topic
-        beta_topic = PublishableContent.objects.get(pk=self.tuto.pk).beta_topic
-        self.assertIsNotNone(TopicAnswerSubscription.objects.get_existing(self.user_author, beta_topic, is_active=True))
-        self.assertEqual(Post.objects.filter(topic=beta_topic).count(), 1)
-        self.assertEqual(beta_topic.tags.count(), 1)
-        self.assertEqual(beta_topic.tags.first().title, sometag.title)
-        # check signal is emitted
-        self.assertEqual(beta_management.send.call_count, 1)
-        self.assertEqual(beta_management.send.call_args[1]["action"], "activate")
-
-        # test if second author follow the topic
-        self.assertIsNotNone(TopicAnswerSubscription.objects.get_existing(second_author, beta_topic, is_active=True))
-        self.assertEqual(TopicRead.objects.filter(topic__pk=beta_topic.pk, user__pk=second_author.pk).count(), 1)
-
-        # test access for public
-        self.client.logout()
-
-        result = self.client.get(
-            reverse("content:view", args=[self.tuto.pk, self.tuto.slug]) + "?version=" + current_sha_beta, follow=False
-        )
-        self.assertEqual(result.status_code, 302)  # (get 302: no access to beta for public)
-
-        # test access for guest;
-        self.client.force_login(self.user_guest)
-
-        # get 200 everywhere :)
-        result = self.client.get(
-            reverse("content:view", args=[tuto.pk, tuto.slug]) + "?version=" + current_sha_beta, follow=False
-        )
-        self.assertEqual(result.status_code, 200)
-
-        result = self.client.get(
-            reverse(
-                "content:view-container", kwargs={"pk": tuto.pk, "slug": tuto.slug, "container_slug": self.part1.slug}
-            )
-            + "?version="
-            + current_sha_beta,
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 200)
-
-        result = self.client.get(
-            reverse(
-                "content:view-container",
-                kwargs={
-                    "pk": tuto.pk,
-                    "slug": tuto.slug,
-                    "parent_container_slug": self.part1.slug,
-                    "container_slug": self.chapter1.slug,
-                },
-            )
-            + "?version="
-            + current_sha_beta,
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 200)
-
-        # change beta version
-        self.client.logout()
-        self.client.force_login(self.user_author)
-
-        result = self.client.post(
-            reverse(
-                "content:edit-container", kwargs={"pk": tuto.pk, "slug": tuto.slug, "container_slug": self.part1.slug}
-            ),
-            {
-                "title": "Un autre titre",
-                "introduction": "Introduire la chose",
-                "conclusion": "Et terminer sur un truc bien",
-                "last_hash": self.part1.compute_hash(),
-            },
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 302)
-
-        tuto = PublishableContent.objects.get(pk=tuto.pk)
-        self.assertNotEqual(current_sha_beta, tuto.sha_draft)
-
-        # add third author:
-        third_author = ProfileFactory().user
-        tuto.authors.add(third_author)
-        tuto.save()
-
-        self.assertIsNone(TopicAnswerSubscription.objects.get_existing(third_author, beta_topic, is_active=True))
-        self.assertEqual(TopicRead.objects.filter(topic__pk=beta_topic.pk, user__pk=third_author.pk).count(), 0)
-
-        # change beta:
-        old_sha_beta = current_sha_beta
-        current_sha_beta = tuto.sha_draft
-        result = self.client.post(
-            reverse("content:set-beta", kwargs={"pk": tuto.pk, "slug": tuto.slug}),
-            {"version": current_sha_beta},
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 302)
-
-        tuto = PublishableContent.objects.get(pk=tuto.pk)
-        self.assertEqual(tuto.sha_beta, current_sha_beta)
-
-        # No new message added since the last time, because the last message
-        # was posted since less than 15 minutes ago.
-
-        self.assertEqual(Post.objects.filter(topic=beta_topic).count(), 1)
-
-        # test if third author follow the topic
-        self.assertIsNotNone(TopicAnswerSubscription.objects.get_existing(third_author, beta_topic, is_active=True))
-        self.assertEqual(TopicRead.objects.filter(topic__pk=beta_topic.pk, user__pk=third_author.pk).count(), 1)
-
-        # then test for guest
-        self.client.logout()
-        self.client.force_login(self.user_guest)
-
-        result = self.client.get(
-            reverse("content:view", args=[tuto.pk, tuto.slug]) + "?version=" + old_sha_beta, follow=False
-        )
-        self.assertEqual(result.status_code, 403)  # no access using the old version
-
-        result = self.client.get(
-            reverse("content:view", args=[tuto.pk, tuto.slug]) + "?version=" + current_sha_beta, follow=False
-        )
-        self.assertEqual(result.status_code, 200)  # ok for the new version
-
-        # inactive beta
-        self.client.logout()
-        self.client.force_login(self.user_author)
-
-        result = self.client.post(
-            reverse("content:inactive-beta", kwargs={"pk": tuto.pk, "slug": tuto.slug}),
-            {"version": current_sha_beta},
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 302)
-
-        self.assertEqual(Post.objects.filter(topic=beta_topic).count(), 2)  # a new message was added !
-        self.assertTrue(Topic.objects.get(pk=beta_topic.pk).is_locked)  # beta was inactived, so topic is locked !
-        # check signal is emitted
-        self.assertEqual(beta_management.send.call_count, 3)
-        self.assertEqual(beta_management.send.call_args[1]["action"], "deactivate")
-
-        # then test for guest
-        self.client.logout()
-        self.client.force_login(self.user_guest)
-
-        result = self.client.get(
-            reverse("content:view", args=[tuto.pk, tuto.slug]) + "?version=" + current_sha_beta, follow=False
-        )
-        self.assertEqual(result.status_code, 403)  # no access anymore
-
-        # reactive beta
-        self.client.logout()
-        self.client.force_login(self.user_author)
-
-        result = self.client.post(
-            reverse("content:set-beta", kwargs={"pk": tuto.pk, "slug": tuto.slug}),
-            {"version": old_sha_beta},  # with a different version than the draft one
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 302)
-
-        tuto = PublishableContent.objects.get(pk=tuto.pk)
-        self.assertEqual(tuto.sha_beta, old_sha_beta)
-
-        # No new message added since the last time, because the last message
-        # was posted since less than 15 minutes ago.
-
-        self.assertEqual(Post.objects.filter(topic=beta_topic).count(), 3)  # a new message was added !
-        self.assertFalse(Topic.objects.get(pk=beta_topic.pk).is_locked)  # not locked anymore
-
-        # then test for guest
-        self.client.logout()
-        self.client.force_login(self.user_guest)
-
-        result = self.client.get(
-            reverse("content:view", args=[tuto.pk, tuto.slug]) + "?version=" + tuto.sha_draft, follow=False
-        )
-        self.assertEqual(result.status_code, 403)  # no access on the non-beta version (of course)
-
-        result = self.client.get(
-            reverse("content:view", args=[tuto.pk, tuto.slug]) + "?version=" + old_sha_beta, follow=False
-        )
-        self.assertEqual(result.status_code, 200)  # access granted
-
-    def test_beta_helps(self):
-        """Check that editorial helps are visible on the beta"""
-
-        # login with author
-        self.client.force_login(self.user_author)
-
-        # create and add help
-        help = HelpWritingFactory()
-        help.save()
-
-        tuto = PublishableContent.objects.get(pk=self.tuto.pk)
-        tuto.helps.add(help)
-        tuto.save()
-
-        # activate beta
-        result = self.client.post(
-            reverse("content:set-beta", kwargs={"pk": tuto.pk, "slug": tuto.slug}),
-            {"version": tuto.sha_draft},
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 302)
-
-        # check that the information is displayed on the beta page
-        result = self.client.get(reverse("content:beta-view", args=[tuto.pk, tuto.slug]), follow=False)
-        self.assertEqual(result.status_code, 200)
-        self.assertContains(result, _("L’auteur de ce contenu recherche"))
-        # and on a container
-        result = self.client.get(
-            reverse(
-                "content:beta-view-container",
-                kwargs={"pk": tuto.pk, "slug": tuto.slug, "container_slug": self.part1.slug},
-            ),
-            follow=False,
-        )
-        self.assertEqual(result.status_code, 200)
-        self.assertContains(result, _("L’auteur de ce contenu recherche"))
 
     def test_history_navigation(self):
         """ensure that, if the title (and so the slug) of the content change, its content remain accessible"""
